+++ src/core/providers/types/providerFactory.types.ts
@@ -1,17 +1,26 @@
 import type { IDAppProviderBase } from '@multiversx/sdk-dapp-utils';
 
+// @ts-ignore
 export interface IProvider extends IDAppProviderBase {
   init: () => Promise<boolean>;
-  // TODO change return type to { address: string, signature: string } and also change the return type in IDAppProviderBase.
-  login: (options?: { token?: string }) => Promise<string | boolean>;
+  login: (options?: { callbackUrl?: string; token?: string }) => Promise<{
+    address: string;
+    signature: string;
+    multisig?: string;
+    impersonate?: string;
+    [key: string]: unknown;
+  }>;
   logout: () => Promise<boolean>;
   setShouldShowConsentPopup?: (shouldShow: boolean) => void;
-  getAddress(): string | undefined;
+  getType: () => ProviderTypeEnum;
+  getAddress(): Promise<string | undefined>;
   // TODO will be removed as soon as the new login method is implemented in the same way for all providers
   getTokenLoginSignature(): string | undefined;
+  // getExtraInfoData(): { multisig?: string; impersonate?: string } | undefined;
 }
 
 export interface IProviderConfig {
+  // TODO check if we have to pass the network object as argument here or it should be read from the state
   network: {
     walletAddress: string;
   };