+++ src/store/selectors/configSelectors.ts
@@ -1,0 +1,6 @@
+import { StoreType } from 'store/store.types';
+
+export const configSelector = ({ config }: StoreType) => config;
+
+export const nativeAuthConfigSelector = ({ config }: StoreType) =>
+  config.nativeAuthConfig;