+++ src/store/slices/loginInfo/index.ts
@@ -1,1 +1,1 @@
-export * from './loginInfoSlice';
+export { loginInfoSlice } from './loginInfoSlice';