+++ src/utils/account/getIsLoggedIn.ts
@@ -1,5 +1,0 @@
-import { getAddress } from './getAddress';
-
-export const getIsLoggedIn = () => {
-  return Boolean(getAddress());
-};